Final changes and adding some readme stuff
