+++ think_big_2023/.env_sample
@@ -1,0 +1,8 @@
+# Sample Environment File, copy this to .env and fill in the blanks
+
+OPENAI_ORG=org-9X9X9X9X9X9X9X9X9X9X9X9X
+OPENAI_API_KEY=sk-9X9X9X9X9X9X9X9X9X9X9X9X
+DISCORD_BOT_TOKEN="9X9X9X9X9X9X9X9X9X9X9X9X"
+
+# This will only need to be set if you are using a custom redis instance
+REDIS_URL=redis://redis:6379
