--- think_big_2023/src/new_main.py
@@ -1,6 +1,0 @@
-from bot import bot, TOKEN
-from tasks import *
-
-
-if __name__ == '__main__':
-    bot.run(TOKEN)
