+++ think_big_2023/src/tasks/__init__.py
@@ -1,3 +1,6 @@
 __all__ = []
 
+# Tasks that are registered will be added to __all__, then imported by main.py
 from tasks.task import *
+from tasks.llm import *
+from tasks.stable_diffusion import *
